Fix duplication of platforms and floors, renamed spawn floors method to be more correct
